src/requirements.txt
requests
openai
fastapi
uvicorn
pydantic
PyJWT
python-dotenv
redis
python-multipart
pytest
pandas